dumb components
those are essentially pure view components, without any knowledge of
redux, actions or effects. Given the correct props, they describe how to
render the react-three-renderer objects on the canvas. Notice that World
is just a group of objects, and even though it doesn't uses it's props
by itself, it serves as an interface for high-order-components to extend
on. That's basically the idea of dumb reusable components vs smart
components.