smart containers
so you compose hocs with dumb components and then you get smart
containers! They are the ones that actually can handle them selves and
express the redux store state into the dumb components. Compose
Namespace and Subscribe to a simple mesh and then boom! you have a
moving square on the screen.