reducers
these are where the application state lives. Reducers should be pure
functions and never mutate on their inputs. This allows neat stuff like
time travelling. The view reducer holds any state that represents what
it is being showed in the screen, that is, the react-three-renderer 3d
objects. WorkReducer holds references to running Web Workers, which are
used for messaging stuff to and from the react View. TaskReducer,
together with taskMiddleware, queues any action dispatched from the
view, in order to play nicely with redux-saga, the place that runs any
side-effects.