tasks and effects
redux-saga allows for a very descritive way of handling effects but it
comes with a price that it waits any sync code to run first. This is
essential for a non-blocking solution and allows for good fps but makes
necessary the use of TaskReducer to queue any actions dispatched from
that sync code (the containers) so that they can be run later, with all
their side effects too. so even though the redux-saga code may look like
it will completelly block the view from rendering, they will actually
run in concurrency, because apparently sync code (view) takes preference
over async code (sagas). That's also exactly why we have a
taskMiddleware to queue any actions from the view, but also why this
middleware doesn't queue any actions putted in the reducers by the
sagas, which are them the actual place from where all actions can be
dispatched and have side-effects.